msm: kgsl: Do not memset pages to zero while adding to pool
Doing a memset to zero while adding a page to pool is not
efficient as the page added to the pool can be returned to
system in case shrinker kicks in. In this scenario, time spent
in zeroing the page is a waste.

Instead of zeroing the page while adding it to pool zero the
page when it is taken from the pool. This helps in reducing
the time taken to free big chunk of memory. Also, allocation
time shouldn't be a problem as zeroing of page anyways happens
during allocation in case it is allocated from system.

Bug: 63937165
Change-Id: I41ab2cb88fb4fd9854d2cc9a45bb60fc7013286a
Signed-off-by: Deepak Kumar <dkumar@codeaurora.org>
Signed-off-by: Raphiel Rollerscaperers <raphielscape@outlook.com>
(cherry picked from commit c5d9ede8fc834d714b65fc0975ed8a69bd855a3d)
(cherry picked from commit fd9e0cb2fe8780097107ddf86513436bcdb3b148)